refactor : InputValidator 삭제
